backend/requirements.txt
fastapi
uvicorn
pydantic
pytest
httpx
redis
requests
python-dotenv
numpy
scikit-learn
packaging
google-genai
e2b-code-interpreter
e2b
PyJWT>=2.8.0
cryptography>=41.0.0
